+++ src/core/rotas.js
@@ -15,6 +15,7 @@ export const ROTAS = [
   { hash: '#/inicio', titulo: 'Início', icone: '⌂', carregar: () => import('../pages/inicio.js').then((m) => m.inicioPage) },
   { hash: '#/mapa', titulo: 'Mapa', icone: '⊕', carregar: () => import('../pages/mapa.js').then((m) => m.mapaPage) },
   { hash: '#/odometro', titulo: 'Trajeto', icone: '↔', carregar: () => import('../pages/odometro.js').then((m) => m.odometroPage) },
+  { hash: '#/trilhas', titulo: 'Trilhas', icone: '≡', carregar: () => import('../pages/trilhas.js').then((m) => m.trilhasPage) },
   { hash: '#/navegacao', titulo: 'Navegação', icone: '⌖', carregar: () => import('../pages/navegacao.js').then((m) => m.navegacaoPage) },
   { hash: '#/bussola', titulo: 'Bússola', icone: '◉', carregar: () => import('../pages/bussola.js').then((m) => m.bussolaPage) },
   { hash: '#/clima', titulo: 'Tempestade', icone: '⚡', carregar: () => import('../pages/clima.js').then((m) => m.climaPage) },